+++ RELEASES.md
@@ -438,3 +438,10 @@
   - lru cache now has getSync and setSync methods
   - lru cache now has values method
   - security users and groups refactor
+  
+7.0.0 2018-05-10
+----------------
+  - allowed set and on paths are more permissive with regards to special characters '(' ')' '&'
+  - set paths are not allowed to contain the * character
+  - server side path checking
+  - increment set functionality and tests